requirements.txt
Django==3.2.18
django-cors-headers==3.14.0
djangorestframework==3.14.0
pre-commit==2.21.0
psycopg2-binary==2.9.6
python-dotenv==0.21.1
PyJWT==2.7.0
boto3==1.26.141